# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Audio Guide 2026 API
  version: 1.13.0
  description: Canonical API for Audio Guide 2026. Includes fail-closed caching rules.
servers:
  - url: https://audiogid-api.vercel.app/v1
    description: Production Server
paths:
...
  /billing/batch-purchase:
    post:
      tags: [Billing]
      summary: Batch Purchase Validation
      operationId: batchPurchase
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchPurchaseRequest' }
      responses:
        '200':
          description: Validated Products
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchPurchaseResponse' }



  # --- Billing (Stores) ---
  /billing/apple/verify:
    post:
      tags: [Billing]
      summary: Verify Apple App Store Receipt
      operationId: verifyAppleReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receipt, product_id, idempotency_key]
              properties:
                receipt: { type: string, description: "Base64 receipt or JWS transaction" }
                product_id: { type: string }
                idempotency_key: { type: string }
                device_anon_id: { type: string }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseVerifyResponse' }
  /billing/google/verify:
    post:
      tags: [Billing]
      summary: Verify Google Play Purchase
      operationId: verifyGooglePurchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [package_name, product_id, purchase_token, idempotency_key]
              properties:
                package_name: { type: string }
                product_id: { type: string }
                purchase_token: { type: string }
                idempotency_key: { type: string }
                device_anon_id: { type: string }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseVerifyResponse' }
  /billing/entitlements:
    get:
      tags: [Billing]
      summary: Get current user entitlements
      operationId: getEntitlements
      parameters:
        - name: device_anon_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Active entitlements
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EntitlementGrantRead' }
  /billing/restore:
    post:
      tags: [Billing]
      summary: Enqueue Restore Purchases (Async)
      operationId: restorePurchases
      description: Initiates server-side reconcile with Apple/Google to recover grants.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idempotency_key, device_anon_id]
              properties:
                platform: { type: string, enum: [apple, google, auto], default: auto }
                idempotency_key: { type: string }
                device_anon_id: { type: string }
                apple_receipt: { type: string, description: "Latest Apple App Store receipt (base64)" }
                google_purchases: 
                  type: array
                  items: { $ref: '#/components/schemas/GooglePurchaseItem' }
                  description: "List of Google purchases to restore (Batch mode)"
                google_purchase_token: { type: string, description: "DEPRECATED: Use google_purchases" }
                product_id: { type: string, description: "DEPRECATED: Use google_purchases" }
                package_name: { type: string, description: "DEPRECATED: Use google_purchases" }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobEnqueueResponse' }
  /billing/restore/{job_id}:
    get:
      tags: [Billing]
      summary: Get Restore Job Status
      operationId: getRestoreJobStatus
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestoreJobRead' }
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    YookassaSignature:
      type: apiKey
      in: header
      name: X-Yookassa-Signature
  schemas:
...
    PurchaseVerifyResponse:
      type: object
      required: [verified, granted, trace_id]
      properties:
        verified: { type: boolean }
        granted: { type: boolean }
        entitlement_grant_id: { type: string, format: uuid, nullable: true }
        order_id: { type: string, nullable: true }
        trace_id: { type: string }
        error: { type: string, nullable: true }
    EntitlementGrantRead:
      type: object
      properties:
        id: { type: string, format: uuid }
        entitlement_slug: { type: string }
        scope: { type: string }
        ref: { type: string }
        granted_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
        is_active: { type: boolean }
    JobEnqueueResponse:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        status: { type: string }
        trace_id: { type: string }
    RestoreJobRead:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [PENDING, RUNNING, COMPLETED, FAILED] }
        result:
          type: object
          properties:
            platform: { type: string }
            grants_created: { type: integer }
            grants_existing: { type: integer }
            grants_total: { type: integer }
            failed_count: { type: integer }
            items: 
              type: array
              items: { $ref: '#/components/schemas/RestoreItemResult' }
          nullable: true
        last_error: { type: string, nullable: true }
        trace_id: { type: string }
...
    GooglePurchaseItem:
      type: object
      required: [purchase_token]
      properties:
        package_name: { type: string }
        product_id: { type: string }
        purchase_token: { type: string }
    RestoreItemResult:
      type: object
      properties:
        product_id: { type: string, nullable: true }
        status: { type: string, enum: [created, existing, failed] }
        error_code: { type: string, nullable: true }
        order_id: { type: string, nullable: true }
        source_ref_hash_prefix: { type: string, nullable: true }
...
    BatchPurchaseRequest:
      type: object
      required: [device_anon_id]
      properties:
        poi_ids: { type: array, items: { type: string, format: uuid } }
        tour_ids: { type: array, items: { type: string, format: uuid } }
        device_anon_id: { type: string }
    BatchPurchaseResponse:
      type: object
      required: [product_ids, already_owned]
      properties:
        product_ids: { type: array, items: { type: string } }
        already_owned: { type: array, items: { type: string } }
...